requirements.txt
# Core scientific computing libraries
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0

# Machine learning libraries
scikit-learn>=1.0.0
optuna>=3.0.0

# Plotting and visualization
matplotlib>=3.4.0
seaborn>=0.11.0

# Configuration management
PyYAML>=6.0

# Excel file handling
openpyxl>=3.0.0
xlrd>=2.0.0

# Progress bars (for optuna)
tqdm>=4.62.0

# Utility libraries
pathlib2>=2.3.6; python_version < "3.6"